feat: move card-image-template directive to directives folder
