+++ apps/angular/projection/src/app/directive/card-image-template.directive.ts
